+++ lib/yt/associations/ids.rb
@@ -4,10 +4,10 @@ module Yt
   module Associations
     # Provides the `has_one :id` method to YouTube resources, which
     # allows to retrieve the id of a resource knowing only its username.
-    # YouTube resources with ids are: channels (and resources of kind channel).
+    # YouTube resources with ids are: resources.
     module Ids
       def id
-        @id ||= ids.first
+        @id ||= ids.first!
       end
 
     private