+++ spec/associations/device_auth/details_sets_spec.rb
@@ -13,7 +13,7 @@ describe Yt::Associations::DetailsSets, scenario: :device_app do
 
     context 'given an unknown video' do
       let(:video) { Yt::Video.new id: 'not-a-video-id', auth: account }
-      it { expect(video.details_set).to be_nil }
+      it { expect{video.details_set}.to raise_error Yt::Errors::NoItems }
     end
   end
 end