+++ spec/associations/device_auth/ids_spec.rb
@@ -5,7 +5,7 @@ describe Yt::Associations::Ids, scenario: :device_app do
   let(:account) { Yt.configuration.account }
   subject(:resource) { Yt::Resource.new url: url, auth: account.auth }
 
-  describe 'id and username' do
+  describe '#id' do
     context 'given a URL containing an existing username' do
       let(:url) { 'youtube.com/fullscreen' }
       it { expect(resource.id).to eq 'UCxO1tY8h1AhOz0T4ENwmpow' }
@@ -13,7 +13,7 @@ describe Yt::Associations::Ids, scenario: :device_app do
 
     context 'given a URL containing an unknown username' do
       let(:url) { 'youtube.com/--not--a--valid--username' }
-      it { expect(resource.id).to be_nil }
+      it { expect{resource.id}.to raise_error Yt::Errors::NoItems }
     end
   end
 end