+++ spec/errors/no_items_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+require 'yt/errors/no_items'
+
+describe Yt::Errors::NoItems do
+  let(:msg) { %r{request.+?returned no items} }
+  describe '#exception' do
+    it { expect{raise Yt::Errors::NoItems}.to raise_error msg }
+  end
+end